Homeworks week-3 pra, week-4 and week-4-pra
Dear Javier,
please find my homeworks revised based on the comments you made to the first batch (in particular the if/else). The one I could not convert to avoid what you mentioned in your revision is the week-3/level 3/practise-all...not sure why...
With regard to the week4 ones I had issues with week4D-Methods exercises 4 and 5, E- exercises 2 and 3, F-exercise part 1. Thank you and have a good afternoon. Flaminia